reordered z3_expression
